backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
sqlalchemy
requests
bcrypt
python-jose[cryptography]
pydantic[email]
twilio
python-dotenv
google-generativeai
elevenlabs
python-socketio
aiofiles
websockets
qdrant-client
openai
cohere
tiktoken
